pf: use M_IP_NEXTHOP and M_IP6_NEXTHOP
Allow pf(4) to use the tagging as well while imposing consistency
inside the pfil hook.  The interface is stored using its index
for the time being as it is lockless and may only end up directing
packets to the wrong interface if it was changed.

Packets through pf_route() and pf_route6() now take the long way
to ip_output() and ip6_output(), respectively, with the interface
switch being enforced there.